+++ flight/tasks/obdh.py
@@ -26,7 +26,7 @@ class Task(TemplateTask):
                 self.SD_scanned = True
 
             # TODO Temporarily start global state switch here
-            # TODO should have a verification checklist in monitor and switch from there
+            # TODO should have a verification checklist in cdh and switch from there
             SM.switch_to("NOMINAL")
         elif SM.current_state == "NOMINAL":
             self.SD_stored_volume = DH.compute_total_size_files()
